+++ src/main/java/cleanie/repatch/post/service/PostService.java
@@ -3,11 +3,13 @@ package cleanie.repatch.post.service;
 import cleanie.repatch.common.exception.BadRequestException;
 import cleanie.repatch.common.exception.model.ExceptionCode;
 import cleanie.repatch.photo.domain.PhotoEntity;
-import cleanie.repatch.photo.service.PostPhotoManager;
+import cleanie.repatch.photo.component.PostPhotoManager;
+import cleanie.repatch.post.component.PostConverter;
+import cleanie.repatch.post.component.PostValidator;
 import cleanie.repatch.post.domain.PostEntity;
-import cleanie.repatch.post.model.PostIdResponse;
-import cleanie.repatch.post.model.PostRequest;
-import cleanie.repatch.post.model.PostResponse;
+import cleanie.repatch.post.model.response.PostIdResponse;
+import cleanie.repatch.post.model.request.PostRequest;
+import cleanie.repatch.post.model.response.PostResponse;
 import cleanie.repatch.post.repository.PostRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
